+++ src/logia/frontend.cpp
@@ -47,7 +47,7 @@ namespace logia
             }
             if (count == line)
             {
-                for (int i = 0; i < column; ++i)
+                for (size_t i = 0; i < column; ++i)
                 {
                     cout << " ";
                 }
